requirements.txt
dnspython==2.7.0
pymongo==4.13.2
python-dotenv==1.1.1
